Ignore .gpx and .xmp files when adding images.
This makes it more convenient to just select all files in a folder and
drop them into the image list.

Note: video files still trigger a warning saying that the image could
not be loaded. I was unsure whether it is better to make it explicit
that videos can not be handled by kgeotag or whether they should be
ignored.